met cov variables as double

+++ NtupleProducer/include/Event.h
@@ -23,10 +23,10 @@ class Event : public Base
         float metpy()                       {return _metpt*sin(_metphi);};
         float metsumet()                    {return _metsumet;};
 
-        float metcov00()		    {return _metcov00;};
-        float metcov01()		    {return _metcov01;};
-        float metcov10()		    {return _metcov10;};
-        float metcov11()		    {return _metcov11;};
+        double metcov00()		    {return _metcov00;};
+        double metcov01()		    {return _metcov01;};
+        double metcov10()		    {return _metcov10;};
+        double metcov11()		    {return _metcov11;};
 
         float metNoHF_pt()                  {return _metNoHF_pt;};
         float metNoHF_phi()                 {return _metNoHF_phi;};
@@ -68,10 +68,10 @@ class Event : public Base
         float _metpt;
         float _metphi;
         float _metsumet;
-	float _metcov00;
-        float _metcov01;
-        float _metcov10;
-        float _metcov11;
+	double _metcov00;
+        double _metcov01;
+        double _metcov10;
+        double _metcov11;
 
         float _metNoHF_pt;
         float _metNoHF_phi;
